Update Exercise_with_solutions.ipynb
Fix the code to match result with what the text asks for. i.e. If text asks to retrieve from 3rd up to 7th (but not including) column to retrieve data from 3rd up and including to 6th column rather than from 4th to 6th, and couple more scenarios corrected.